refactor: move tls to a separate file
