+++ sztp-agent/pkg/secureagent/utils.go
@@ -10,17 +10,10 @@ package secureagent
 
 import (
 	"bufio"
-	"bytes"
-	"crypto/tls"
-	"crypto/x509"
 	"encoding/json"
-	"errors"
-	"io"
 	"log"
-	"net/http"
 	"os"
 	"regexp"
-	"strconv"
 	"strings"
 
 	"github.com/jaypipes/ghw"
@@ -47,77 +40,6 @@ func extractfromLine(line, regex string, index int) string {
 		return ""
 	}
 	return re.FindAllString(line, -1)[index]
-}
-
-func (a *Agent) doTLSRequest(input string, url string, empty bool) (*BootstrapServerPostOutput, error) {
-	var postResponse BootstrapServerPostOutput
-	var errorResponse BootstrapServerErrorOutput
-
-	log.Println("[DEBUG] Sending to: " + url)
-	log.Println("[DEBUG] Sending input: " + input)
-
-	body := strings.NewReader(input)
-	r, err := http.NewRequest(http.MethodPost, url, body)
-	if err != nil {
-		return nil, err
-	}
-
-	r.SetBasicAuth(a.GetSerialNumber(), a.GetDevicePassword())
-	r.Header.Add("Content-Type", a.GetContentTypeReq())
-
-	caCert, _ := os.ReadFile(a.GetBootstrapTrustAnchorCert())
-	caCertPool := x509.NewCertPool()
-	caCertPool.AppendCertsFromPEM(caCert)
-	cert, _ := tls.LoadX509KeyPair(a.GetDeviceEndEntityCert(), a.GetDevicePrivateKey())
-
-	client := &http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{ //nolint:gosec
-				RootCAs:      caCertPool,
-				Certificates: []tls.Certificate{cert},
-			},
-		},
-	}
-	res, err := client.Do(r)
-	if err != nil {
-		log.Println("Error doing the request", err.Error())
-		return nil, err
-	}
-	defer func() {
-		if err := res.Body.Close(); err != nil {
-			log.Println("Error when closing:", err)
-		}
-	}()
-
-	bodyBytes, err := io.ReadAll(res.Body)
-	if err != nil {
-		log.Println("Error reading the request", err.Error())
-		return nil, err
-	}
-
-	decoder := json.NewDecoder(bytes.NewReader(bodyBytes))
-	decoder.DisallowUnknownFields()
-	if !empty {
-		derr := decoder.Decode(&postResponse)
-		if derr != nil {
-			errdecoder := json.NewDecoder(bytes.NewReader(bodyBytes))
-			errdecoder.DisallowUnknownFields()
-			eerr := errdecoder.Decode(&errorResponse)
-			if eerr != nil {
-				log.Println("Received unknown response", string(bodyBytes))
-				return nil, derr
-			}
-			return nil, errors.New("[ERROR] Expected conveyed-information" +
-				", received error type=" + errorResponse.IetfRestconfErrors.Error[0].ErrorType +
-				", tag=" + errorResponse.IetfRestconfErrors.Error[0].ErrorTag +
-				", message=" + errorResponse.IetfRestconfErrors.Error[0].ErrorMessage)
-		}
-		log.Println(postResponse)
-	}
-	if res.StatusCode != http.StatusOK {
-		return nil, errors.New("[ERROR] Status code received: " + strconv.Itoa(res.StatusCode) + " ...but status code expected: " + strconv.Itoa(http.StatusOK))
-	}
-	return &postResponse, nil
 }
 
 // GetSerialNumber returns the serial number of the device
