+++ sztp-agent/pkg/secureagent/utils_test.go
@@ -5,53 +5,9 @@
 package secureagent
 
 import (
-	"reflect"
 	"strings"
 	"testing"
 )
-
-func TestAgent_doTLSRequest(t *testing.T) {
-	type fields struct {
-		BootstrapURL             string
-		SerialNumber             string
-		DevicePassword           string
-		DevicePrivateKey         string
-		DeviceEndEntityCert      string
-		BootstrapTrustAnchorCert string
-		ContentTypeReq           string
-		InputJSONContent         string
-		DhcpLeaseFile            string
-	}
-	var tests []struct {
-		name    string
-		fields  fields
-		want    *BootstrapServerPostOutput
-		wantErr bool
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			a := &Agent{
-				BootstrapURL:             tt.fields.BootstrapURL,
-				SerialNumber:             tt.fields.SerialNumber,
-				DevicePassword:           tt.fields.DevicePassword,
-				DevicePrivateKey:         tt.fields.DevicePrivateKey,
-				DeviceEndEntityCert:      tt.fields.DeviceEndEntityCert,
-				BootstrapTrustAnchorCert: tt.fields.BootstrapTrustAnchorCert,
-				ContentTypeReq:           tt.fields.ContentTypeReq,
-				InputJSONContent:         tt.fields.InputJSONContent,
-				DhcpLeaseFile:            tt.fields.DhcpLeaseFile,
-			}
-			got, err := a.doTLSRequest(a.GetInputJSONContent(), a.GetBootstrapURL(), false)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("doTLSRequest() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("doTLSRequest() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
 
 func Test_extractfromLine(t *testing.T) {
 	type args struct {
